cheri: fs: afs: Fix uintptr_t usage in fs/afs/ and rxrpc
Fix uintptr_t usage in fs/afs/ and the related net/rxrpc.

Use uintptr_t for values that can contain a pointer in fs/afs/
and cast as necessary:
- The ->d_fsdata field stores the data version which is not a pointer.
- The procfs callbacks pass the file offset as the iterator.
- The call_user_ID parameter to verious functions must be able to hold
  a pointer.

+++ fs/afs/cm_security.c
@@ -33,7 +33,7 @@ static int afs_respond_to_challenge(struct sk_buff *challenge)
 	struct afs_server *server;
 #endif
 	struct rxrpc_peer *peer;
-	unsigned long peer_data;
+	uintptr_t peer_data;
 	u16 service_id;
 	u8 security_index;
